delete worker

+++ src/pages/list.jsx
@@ -5,6 +5,11 @@ import Swal from "sweetalert2"
 import { useNavigate } from "react-router-dom"
 
 const List = () => {
+    const navigate = useNavigate()
+
+    const editEmp = (id) => {
+        navigate(`${id}`)
+    }
     const list = useSelector(s => s.employees)
     return (
         <>
@@ -19,11 +24,11 @@ const List = () => {
                     </TableRow>
                 </TableHeader>
                 <TableBody>
-                    {list.map((emp) => 
-                        <TableRow>
+                    {list.map((emp) =>
+                        <TableRow key={emp.id}>
                             <TableCell>
-                                <Button icon='edit' onClick={()=>deleteEmp(emp.id)} />
-                                <Button icon='trash' onClick={()=>editEmp(emp.id)}/>
+                                <Button icon='trash' onClick={() => deleteEmp(emp.id)} />
+                                <Button icon='edit' onClick={() => editEmp(emp.id)} />
                             </TableCell>
                             <TableCell>{emp.dateStart}</TableCell>
                             <TableCell>{emp.firstName}</TableCell>
@@ -38,25 +43,28 @@ const List = () => {
 }
 export default List;
 
-const deleteEmp = (id)=>{
-    employees.deleteEmployee(id).then(res=>{
-        Swal.fire({
-            icon: "success",
-            showConfirmButton: false,
-            title: 'נמחק בהצלחה',
-            timer:1500
-        })
-    }).catch(err=>{
-        Swal.fire({
-            icon: 'error',
-            showConfirmButton:false,
-            timer: 1500,
-            text: err.message
-        })
+const deleteEmp = (id) => {
+    Swal.fire({
+        showConfirmButton: true,
+        showCloseButton: true,
+        icon: 'question',
+        text: 'Are you sure?'
+    }).then(res => {
+        if (res.isConfirmed)
+            employees.deleteEmployee(id).then(res => {
+                Swal.fire({
+                    icon: "success",
+                    showConfirmButton: false,
+                    title: 'נמחק בהצלחה',
+                    timer: 1500
+                })
+            }).catch(err => {
+                Swal.fire({
+                    icon: 'error',
+                    showConfirmButton: false,
+                    timer: 1500,
+                    text: err.message
+                })
+            })
     })
-}
-
-const editEmp = (id)=>{
-    const navigate = useNavigate()
-    navigate(`/${id}`)
 }
